Fix relative paths bug for SSR
In your frontendmasters SSR chapter, last video (00:03:58) you encounter a bug. You get a `index.html` version of the `/details:id` page, but with no styling. That's because when you refreshed the page it made a request to the static resources, but with a relative path (`localhost:3000/details/dist/ClientApp.js`). Since it didn't match the `/dist` route of Express, it will always match the default (React render/stream one).

In this commit I just replace the `"dist` with `"/dist` so it will always point to the correct static resource.